+++ example/lib/nodes/simple_qa_node.dart
@@ -8,7 +8,7 @@ import 'show_node_config_dialog.dart';
 
 class SimpleQAConfigWidget extends StatefulWidget {
   const SimpleQAConfigWidget({super.key, this.data});
-  final Map<String, Object>? data;
+  final Map<String, dynamic>? data;
 
   @override
   State<SimpleQAConfigWidget> createState() => _SimpleQAConfigWidgetState();
@@ -85,9 +85,9 @@ class SimpleQAWidget extends StatefulWidget {
 }
 
 class _SimpleQAWidgetState extends State<SimpleQAWidget> {
-  String input = "";
-  String output = "";
-  String prompt = "";
+  late String input = widget.node.data?["input"] as String? ?? "";
+  late String output = widget.node.data?["output"] as String? ?? "";
+  late String prompt = widget.node.data?["prompt"] as String? ?? "";
 
   @override
   Widget build(BuildContext context) {
@@ -106,6 +106,12 @@ class _SimpleQAWidgetState extends State<SimpleQAWidget> {
               output = v["output"]?.toString() ?? "";
               prompt = v["prompt"]?.toString() ?? "";
             });
+
+            widget.node.data = {
+              "input": input,
+              "output": output,
+              "prompt": prompt,
+            };
           }
         });
       },
@@ -196,21 +202,49 @@ class _SimpleQAWidgetState extends State<SimpleQAWidget> {
 }
 
 class SimpleQaNode extends INode {
-  SimpleQaNode({
-    required super.label,
-    required super.uuid,
-    required super.depth,
-    required super.offset,
-    super.children,
-    super.height = 200,
-    super.width = 300,
-    super.nodeName = "单输入输出问答节点",
-    super.description = "仅支持一个输入和一个输出的节点",
-    super.builder,
-  }) {
+  SimpleQaNode(
+      {required super.label,
+      required super.uuid,
+      required super.depth,
+      required super.offset,
+      super.children,
+      super.height = 200,
+      super.width = 200,
+      super.nodeName = "单输入输出问答节点",
+      super.description = "仅支持一个输入和一个输出的节点",
+      super.builder,
+      super.builderName = "SimpleQaNode",
+      super.data}) {
     builder = (context) => SimpleQAWidget(
           node: this,
         );
+  }
+
+  factory SimpleQaNode.fromJson(Map<String, dynamic> json) {
+    String uuid = json["uuid"] ?? "";
+    String label = json["label"] ?? "";
+    int depth = json["depth"] ?? 0;
+    Offset offset = Offset(json["offset"]["dx"], json["offset"]["dy"]);
+    double width = json["width"] ?? 300;
+    double height = json["height"] ?? 400;
+    String nodeName = json["nodeName"] ?? "base";
+    String description =
+        json["description"] ?? "Base node, just for testing purposes";
+    String builderName = json["builderName"] ?? "base";
+    Map<String, dynamic>? data = json["data"];
+
+    return SimpleQaNode(
+      offset: offset,
+      width: width,
+      height: height,
+      nodeName: nodeName,
+      description: description,
+      builderName: builderName,
+      label: label,
+      uuid: uuid,
+      depth: depth,
+      data: data,
+    );
   }
 
   @override