+++ lib/src/nodes/nodes.dart
@@ -1,4 +1,4 @@
-export './base_node.dart';
+export 'edge.dart';
 export './fishbone.dart';
 export './node_widget.dart';
 export './inode.dart';